refactor: CreateLogPage 인라인 컴포넌트 및 사진 업로드 로직 분리
