+++ apps/web/src/views/log/ui/CreateLogPage.tsx
@@ -1,21 +1,18 @@
 'use client';
 
-import { type ChangeEvent, useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 
-import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 
 import {
   Button,
   type DateValue,
   Field,
-  Icon,
   Input,
   Switch,
   Textarea,
   type TimeValue,
 } from '@plog/ui';
-import { cn } from '@plog/utils';
 
 import { PlaceCategorySheet } from '@/features/select-place-category';
 import { ReviewTagsSheet } from '@/features/select-review-tags';
@@ -24,252 +21,13 @@ import { WorkTimeDialog } from '@/features/select-work-time';
 
 import { type PlaceCategoryValue, PlaceTagValue } from '@/entities/place';
 
-import FocusLevelDefault1 from '@/shared/assets/focus-levels/focus-level-default-1.svg';
-import FocusLevelDefault2 from '@/shared/assets/focus-levels/focus-level-default-2.svg';
-import FocusLevelDefault3 from '@/shared/assets/focus-levels/focus-level-default-3.svg';
-import FocusLevelDefault4 from '@/shared/assets/focus-levels/focus-level-default-4.svg';
-import FocusLevelDefault5 from '@/shared/assets/focus-levels/focus-level-default-5.svg';
-import FocusLevelSelect1 from '@/shared/assets/focus-levels/focus-level-select-1.svg';
-import FocusLevelSelect2 from '@/shared/assets/focus-levels/focus-level-select-2.svg';
-import FocusLevelSelect3 from '@/shared/assets/focus-levels/focus-level-select-3.svg';
-import FocusLevelSelect4 from '@/shared/assets/focus-levels/focus-level-select-4.svg';
-import FocusLevelSelect5 from '@/shared/assets/focus-levels/focus-level-select-5.svg';
-
-type FocusLevel = 1 | 2 | 3 | 4 | 5;
-
-type PhotoPreview = {
-  id: string;
-  file: File;
-  url: string;
-};
-
-const MAX_PHOTO_COUNT = 5;
-
-const FOCUS_LEVEL_OPTIONS = [
-  {
-    value: 1,
-    label: '매우 낮음',
-    DefaultIcon: FocusLevelDefault1,
-    SelectedIcon: FocusLevelSelect1,
-  },
-  {
-    value: 2,
-    label: '낮음',
-    DefaultIcon: FocusLevelDefault2,
-    SelectedIcon: FocusLevelSelect2,
-  },
-  {
-    value: 3,
-    label: '보통',
-    DefaultIcon: FocusLevelDefault3,
-    SelectedIcon: FocusLevelSelect3,
-  },
-  {
-    value: 4,
-    label: '높음',
-    DefaultIcon: FocusLevelDefault4,
-    SelectedIcon: FocusLevelSelect4,
-  },
-  {
-    value: 5,
-    label: '매우 높음',
-    DefaultIcon: FocusLevelDefault5,
-    SelectedIcon: FocusLevelSelect5,
-  },
-] as const;
-
-const PRIVACY_SETTING_OPTIONS = [
-  {
-    type: 'all',
-    title: '이 기록은 피드에 공유됩니다',
-    content: '다른 사용자들이 회원님의 환경 기록을 볼 수 있습니다.',
-  },
-  {
-    type: 'private',
-    title: '이 기록은 나만 볼 수 있습니다',
-    content: '비공개로 설정되어 다른 사용자들이 볼 수 없습니다.',
-  },
-] as const;
-
-function createPhotoPreview(file: File, index: number): PhotoPreview {
-  return {
-    id: `${file.name}-${file.size}-${file.lastModified}-${Date.now()}-${index}`,
-    file,
-    url: URL.createObjectURL(file),
-  };
-}
-
-function PhotoUploader({
-  photos,
-  onAdd,
-  onRemove,
-}: {
-  photos: PhotoPreview[];
-  onAdd: (files: File[]) => void;
-  onRemove: (id: string) => void;
-}) {
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const canAddMore = photos.length < MAX_PHOTO_COUNT;
-
-  useEffect(() => {
-    if (!fileInputRef.current) return;
-
-    const dataTransfer = new DataTransfer();
-    photos.forEach(({ file }) => dataTransfer.items.add(file));
-    fileInputRef.current.files = dataTransfer.files;
-  }, [photos]);
-
-  const handleFileChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const ACCEPTED_TYPES = new Set([
-      'image/jpeg',
-      'image/png',
-      'image/heic',
-      'image/heif',
-    ]);
-    const selectedFiles = Array.from(event.target.files ?? []).filter((file) =>
-      ACCEPTED_TYPES.has(file.type),
-    );
-
-    if (selectedFiles.length === 0) return;
-    onAdd(selectedFiles);
-  };
-
-  return (
-    <div className="flex gap-4 pt-1 pb-1">
-      <input
-        ref={fileInputRef}
-        type="file"
-        name="photos"
-        accept=".jpg, .png, .heic"
-        multiple
-        className="sr-only"
-        onChange={handleFileChange}
-      />
-
-      <button
-        type="button"
-        disabled={!canAddMore}
-        className="flex size-25 shrink-0 flex-col items-center justify-center gap-1 rounded-xl border border-semantic-stroke-subtle bg-semantic-bg-standard text-semantic-object-normal transition-colors hover:bg-semantic-bg-deep focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-semantic-stroke-subtle disabled:cursor-not-allowed disabled:text-semantic-object-subtle"
-        aria-label="사진 등록"
-        onClick={() => fileInputRef.current?.click()}
-      >
-        <Icon name="camera-filled" />
-        <span className="label-sm">
-          {photos.length}/{MAX_PHOTO_COUNT}
-        </span>
-      </button>
-
-      <div className="min-w-0 flex-1 overflow-x-auto [-ms-overflow-style:none] [scrollbar-width:none] [&::-webkit-scrollbar]:hidden">
-        <div className="flex w-max gap-4">
-          {photos.map((photo, index) => (
-            <div
-              key={photo.id}
-              className="relative size-25 shrink-0 overflow-hidden rounded-xl bg-semantic-object-subtler"
-            >
-              <Image
-                src={photo.url}
-                alt={`등록된 사진 ${index + 1}`}
-                fill
-                sizes="100px"
-                unoptimized
-                className="object-cover"
-              />
-              <button
-                type="button"
-                className="absolute top-2 right-2 flex size-6 cursor-pointer items-center justify-center rounded-full border-2 border-semantic-system-white bg-semantic-feedback-error-normal text-xl leading-none shadow-[0_2px_6px_rgba(0,0,0,0.16)]"
-                aria-label={`등록된 사진 ${index + 1} 삭제`}
-                onClick={() => onRemove(photo.id)}
-              >
-                <Icon
-                  name="close"
-                  className="text-semantic-system-white"
-                  size={16}
-                />
-              </button>
-            </div>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-}
+import { usePhotoUpload } from '../model/use-photo-upload';
+import PhotoUploader from './PhotoUploader';
+import PrivacySettingSection from './PrivacySettingSection';
+import RatingPicker, { type FocusLevel } from './RatingPicker';
 
 function SectionDivider() {
   return <div className="h-2 bg-semantic-bg-deep" />;
-}
-
-function RatingPicker({
-  value,
-  onChange,
-}: {
-  value: FocusLevel | null;
-  onChange: (score: FocusLevel) => void;
-}) {
-  return (
-    <div className="grid w-full grid-cols-5 gap-2">
-      {FOCUS_LEVEL_OPTIONS.map(
-        ({ value: score, label, DefaultIcon, SelectedIcon }) => {
-          const isSelected = value === score;
-          const RatingIcon = isSelected ? SelectedIcon : DefaultIcon;
-
-          return (
-            <label
-              key={score}
-              className="group relative flex aspect-square cursor-pointer items-center justify-center overflow-visible rounded-sm transition outline-none focus-within:ring-2 focus-within:ring-semantic-accent-normal focus-within:ring-offset-2 focus-within:ring-offset-semantic-bg-standard"
-            >
-              <input
-                type="radio"
-                name="focusScore"
-                value={score}
-                checked={isSelected}
-                onChange={() => onChange(score)}
-                className="sr-only"
-                aria-label={`집중도 ${score}점, ${label}`}
-              />
-              <RatingIcon
-                aria-hidden="true"
-                className="block size-full scale-140"
-              />
-            </label>
-          );
-        },
-      )}
-    </div>
-  );
-}
-
-function PrivacySettingSection({ isPublic }: { isPublic: boolean }) {
-  const privacyType = isPublic ? 'all' : 'private';
-  const { title, content } =
-    PRIVACY_SETTING_OPTIONS.find(({ type }) => type === privacyType) ??
-    PRIVACY_SETTING_OPTIONS[0];
-
-  return (
-    <div
-      className={cn(
-        'flex gap-2 rounded-xl p-4',
-        isPublic ? 'bg-semantic-accent-subtler' : 'bg-semantic-object-subtler',
-      )}
-    >
-      <div
-        className={cn(
-          'shrink-0',
-          isPublic
-            ? 'text-semantic-accent-normal'
-            : 'text-semantic-object-normal',
-        )}
-      >
-        <Icon
-          name="circle-exclamation"
-          className={`${isPublic ? 'text-semantic-accent-normal' : 'text-semantic-object-normal'}`}
-        />
-      </div>
-      <div className="min-w-0 flex-1">
-        <p className="label-lg text-semantic-object-bold">{title}</p>
-        <p className="label-md mt-1 text-semantic-object-normal">{content}</p>
-      </div>
-    </div>
-  );
 }
 
 type CreateFeedPageProps = {
@@ -279,12 +37,9 @@ type CreateFeedPageProps = {
 export default function CreateLogPage({
   initialPlaceName = '',
 }: CreateFeedPageProps) {
-  const photoPreviewsRef = useRef<PhotoPreview[]>([]);
-  const router = useRouter();
   const [title, setTitle] = useState('');
   const [content, setContent] = useState('');
   const [placeName, setPlaceName] = useState(initialPlaceName);
-  const [photos, setPhotos] = useState<PhotoPreview[]>([]);
   const [focusScore, setFocusScore] = useState<FocusLevel | null>(null);
   const [placeCategory, setPlaceCategory] = useState<PlaceCategoryValue | null>(
     null,
@@ -295,36 +50,9 @@ export default function CreateLogPage({
   const [reviewTags, setReviewTags] = useState<PlaceTagValue[]>([]);
   const [isPublic, setIsPublic] = useState(false);
 
-  useEffect(() => {
-    photoPreviewsRef.current = photos;
-  }, [photos]);
+  const router = useRouter();
 
-  useEffect(() => {
-    return () => {
-      photoPreviewsRef.current.forEach(({ url }) => URL.revokeObjectURL(url));
-    };
-  }, []);
-
-  const handleAddPhotos = (files: File[]) => {
-    setPhotos((currentPhotos) => {
-      const availableCount = MAX_PHOTO_COUNT - currentPhotos.length;
-      const nextFiles = files.slice(0, availableCount);
-
-      return [
-        ...currentPhotos,
-        ...nextFiles.map((file, index) => createPhotoPreview(file, index)),
-      ];
-    });
-  };
-
-  const handleRemovePhoto = (id: string) => {
-    setPhotos((currentPhotos) => {
-      const targetPhoto = currentPhotos.find((photo) => photo.id === id);
-      if (targetPhoto) URL.revokeObjectURL(targetPhoto.url);
-
-      return currentPhotos.filter((photo) => photo.id !== id);
-    });
-  };
+  const { photos, handleAddPhotos, handleRemovePhoto } = usePhotoUpload();
 
   const handleClearPlaceName = () => {
     setPlaceName('');
@@ -341,7 +69,6 @@ export default function CreateLogPage({
             onRemove={handleRemovePhoto}
           />
         </Field>
-
         <Field label="제목" required>
           <Input
             value={title}
@@ -354,7 +81,6 @@ export default function CreateLogPage({
             {title.length}/20자
           </span>
         </Field>
-
         <Field label="환경 기록을 작성해 주세요" required>
           <Textarea
             value={content}
@@ -367,9 +93,7 @@ export default function CreateLogPage({
           />
         </Field>
       </section>
-
       <SectionDivider />
-
       <section className="flex flex-col gap-6 px-6 py-6">
         <div className="flex flex-col gap-3">
           <Field label="작업 장소" required>
@@ -381,18 +105,15 @@ export default function CreateLogPage({
               onClick={() => router.push('/log/place-search')}
             />
           </Field>
-
           <PlaceCategorySheet
             value={placeCategory}
             onChange={setPlaceCategory}
           />
         </div>
-
         <div className="flex flex-col gap-3">
           <Field label="작업 날짜" required>
             <WorkDateDialog value={workDate} onChange={setWorkDate} />
           </Field>
-
           <div className="grid grid-cols-2 gap-4">
             <Field label="시작 시간" required>
               <WorkTimeDialog
@@ -413,23 +134,19 @@ export default function CreateLogPage({
           </div>
         </div>
       </section>
-
       <SectionDivider />
-
       <section className="flex flex-col gap-6 px-6 py-6">
         <div className="flex flex-col gap-4">
           <Field label="집중도를 평가해 주세요" required>
             <RatingPicker value={focusScore} onChange={setFocusScore} />
           </Field>
         </div>
-
         <div className="flex flex-col gap-4 border-b border-semantic-stroke-subtler pb-6">
           <Field label="후기 요약 태그를 선택해주세요" required>
             <ReviewTagsSheet value={reviewTags} onChange={setReviewTags} />
           </Field>
         </div>
       </section>
-
       <section className="flex flex-col gap-4 px-6 pt-6 pb-10">
         <Field
           label="공개 설정"
